Moved GraphNode to node package
